Session filtering/ column filters (#9930)
* add column filters

* add test

* fix toggle for filter

* small tweak
