+++ services/QuillLMS/client/app/bundles/Shared/interfaces/misc.ts
@@ -2,3 +2,12 @@ export interface UserIdsForEvent {
   studentId: number,
   teacherId: number
 }
+
+export interface NumberFilterInputProps {
+  handleChange: Function;
+  label: string;
+  column: {
+    filterValue: string;
+    id: string;
+  }
+}
