+++ services/QuillLMS/client/app/constants/evidence.ts
@@ -293,61 +293,3 @@ export const numericalWordOptions = {
   3: 'Fourth',
   4: 'Fifth'
 }
-
-export const activitySessionIndexResponseHeaders = [
-  {
-    Header: "Date | Time",
-    accessor: "datetime",
-    width: 150
-  },
-  {
-    Header: "Session ID",
-    accessor: "session_uid",
-    width: 100
-  },
-  {
-    Header: "Total Responses",
-    accessor: "total_attempts",
-    width: 150
-  },
-  {
-    Header: "Because",
-    accessor: "because_attempts",
-    width: 100
-  },
-  {
-    Header: "But",
-    accessor: "but_attempts",
-    width: 100
-  },
-  {
-    Header: "So",
-    accessor: "so_attempts",
-    width: 100
-  },
-  {
-    Header: "Scored",
-    accessor: "scored_count",
-    width: 100
-  },
-  {
-    Header: "Weak",
-    accessor: "weak_count",
-    width: 100
-  },
-  {
-    Header: "Strong",
-    accessor: "strong_count",
-    width: 100
-  },
-  {
-    Header: "Completed?",
-    accessor: "completed",
-    width: 100
-  },
-  {
-    Header: "",
-    accessor: "view_link",
-    width: 100
-  },
-]
